// src/cli/types.ts
import { LanguageCode } from '../data/languageMap';

export interface CliOptions {
  defaultLocale?: string;
  targetLocales?: string[];
  outputDir?: string;
  apiKey?: string;
  apiUrl?: string;
}

export interface QuestionnaireAnswers {
  defaultLocale: string;
  targetLocales: string[];
  outputDir: string;
  apiKey?: string;
  apiUrl?: string;
}

export interface InitConfig {
  defaultLocale: string;
  targetLocales: string[];
  outputDir: string;
  apiKey?: string;
  apiUrl?: string;
}